+++ server/app/middlewares/auth0.js
@@ -1,0 +1,23 @@
+// authentication middlewares
+const jwt = require('express-jwt');
+const jwks = require('jwks-rsa');
+const jwtAuthz = require('express-jwt-authz');
+
+
+var jwtCheck = jwt({
+  secret: jwks.expressJwtSecret({
+    cache: true,
+    rateLimit: true,
+    jwksRequestsPerMinute: 5,
+    jwksUri: `https://${process.env.AUTH0_DOMAIN}/.well-known/jwks.json`
+  }),
+  audience: `https://${process.env.AUTH0_AUDIENCE}/`,
+  issuer: `https://${process.env.AUTH0_DOMAIN}/`,
+  algorithms: ['RS256']
+});
+
+
+module.exports = {
+  jwtCheck,
+  jwtAuthz
+}
